requirements.txt
alembic==1.13.0
appdirs==1.4.4
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
certifi==2023.11.17
click==8.1.7
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
gunicorn==21.2.0
importlib-metadata==7.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
pyee==8.2.2
pyppeteer==1.0.2
pytest==7.4.3
python-dotenv==1.0.0
python-http-client==3.3.7
sendgrid==6.11.0
soupsieve==2.5
SQLAlchemy==2.0.23
starkbank-ecdsa==2.2.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==1.26.18
websockets==10.4
Werkzeug==3.0.1
zipp==3.17.0
